Fix how I grab version numbers
There seems to be bugs in the gem provider that will return extra words in the version string.  Also, the apt provider will return package instances that are not installed, but have config files on the system so are not purged.  This caused how I grab the version numbers out of the reports to fail sometimes.