Fix current version parsing
The gem provider will throw extra words in the report output for the current version of an instance, so I had to work around that.  Also, I had to work around an issue where apt will report a package as installed when you get a list of instances, but the package's state is reported as absent.  These are packages that have been removed, but not purged.