Use /opt/puppetlabs/facter/facts.d directory
Previous to this commit, the /etc/puppetlabs/facter/facts.d directory
was used to cache the package_updates Facter fact. The Puppet Enterprise
installer did not create that directory, forcing this module to ensure
its existence.  However, since the directory might be managed elsewhere,
this module used a nasty exec to ensure the directory existed.

It turns out this was not necessary as the Puppet Enterprise (and open
source) packages create a /opt/puppetlabs/facter/facts.d directory that
Facter defaults to searching for custom facts.

This commit uses the /opt/puppetlabs/facter/facts.d directory instead of
the /etc/puppetlabs/facter/facts.d directory to remove the the evil exec