Do not override versions specified in Puppet code
Previous to this commit, versions specified in Puppet code would be
overriden if the package has an update specified.  Puppet code should
*always* take precedence, so this commit tests if the version is being
managed by Puppet code.

If the version is being managed and there is an update specified that is
different, a warning is output as part of the report. However, that
warning will not show up in Event Inspection