Restructure package hash and clean up find code
This commit introduces a Puppet::Node::Patches#find method that calls
the indirector find method itself. The two existing termini had
duplicate code that can now live in this method

Also, the package_updates hash has a new key called 'classes' that's
value can be a hash where Puppet classes can be the key and the value is
hash with version and provider keys. Any package versions specified
outside of the classes key are considered global.

For example:

---
package_updates:
  classes:
    role::webserver
      apache:
        version: '2.9.3.el7'
        provider: 'yum'
   gcc:
     version: '4.8.5-4.el7'
     provider: 'yum'